Big bad reformat

+++ modules/library/xml/src/test/java/org/geotools/xml/XMLParserTest.java
@@ -1,7 +1,7 @@
 /*
  *    GeoTools - The Open Source Java GIS Toolkit
  *    http://geotools.org
- * 
+ *
  *    (C) 2004-2008, Open Source Geospatial Foundation (OSGeo)
  *
  *    This library is free software; you can redistribute it and/or
@@ -16,36 +16,28 @@
  */
 package org.geotools.xml;
 
+import static junit.framework.TestCase.assertNotNull;
+
 import java.io.File;
 import java.net.URI;
 import java.util.logging.Level;
-
 import javax.naming.OperationNotSupportedException;
 import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
-
 import junit.framework.TestCase;
-import static junit.framework.TestCase.assertNotNull;
-
 import org.geotools.TestData;
 import org.geotools.xml.schema.Schema;
 import org.xml.sax.SAXException;
 
-
 /**
- * <p>
- * DOCUMENT ME!
- * </p>
- * @
+ * DOCUMENT ME! @
  *
  * @author dzwiers www.refractions.net
- *
- *
  * @source $URL$
  */
 public class XMLParserTest extends TestCase {
-    
-     public void testNestedFeature() throws Throwable {
+
+    public void testNestedFeature() throws Throwable {
         SAXParserFactory spf = SAXParserFactory.newInstance();
         spf.setNamespaceAware(true);
         spf.setValidating(false);
@@ -53,7 +45,7 @@ public class XMLParserTest extends TestCase {
         SAXParser parser = spf.newSAXParser();
 
         String path = "/wfsgetfeature.xml";
-        File f = TestData.copy(this,path);
+        File f = TestData.copy(this, path);
         XMLSAXHandler xmlContentHandler = new XMLSAXHandler(null);
         XMLSAXHandler.setLogLevel(Level.FINEST);
         XSISAXHandler.setLogLevel(Level.FINEST);
@@ -65,16 +57,17 @@ public class XMLParserTest extends TestCase {
         Object doc = xmlContentHandler.getDocument();
         assertNotNull("Document missing", doc);
     }
-    public void testMail(){
-        try {            
+
+    public void testMail() {
+        try {
             String path = "xml/mails.xml";
 
-            File f = TestData.copy(this,path);
-            TestData.copy(this,"xml/mails.xsd");
+            File f = TestData.copy(this, path);
+            TestData.copy(this, "xml/mails.xsd");
             URI u = f.toURI();
 
-            Object doc = DocumentFactory.getInstance(u,null,Level.WARNING);
-            
+            Object doc = DocumentFactory.getInstance(u, null, Level.WARNING);
+
             assertNotNull("Document missing", doc);
             System.out.println(doc);
         } catch (SAXException e) {
@@ -85,40 +78,39 @@ public class XMLParserTest extends TestCase {
             fail(e.toString());
         }
     }
-    
-    public void testMailWrite(){
 
-        try {            
-        String path = "xml/mails.xml";
+    public void testMailWrite() {
 
-        File f = TestData.copy(this,path);
-        TestData.copy(this,"xml/mails.xsd");
+        try {
+            String path = "xml/mails.xml";
 
-        Object doc = DocumentFactory.getInstance(f.toURI(),null,Level.WARNING);
-        assertNotNull("Document missing", doc);
+            File f = TestData.copy(this, path);
+            TestData.copy(this, "xml/mails.xsd");
 
-        Schema s = SchemaFactory.getInstance(new URI("http://mails/refractions/net"));
-                
-        path = "mails_out.xml";
-        f = TestData.temp(this, path);
-        if(f.exists())
-            f.delete();
-        f.createNewFile();
-        
-        DocumentWriter.writeDocument(doc,s,f,null);
-        
-        doc = DocumentFactory.getInstance(f.toURI(),null,Level.WARNING);
-        assertNotNull("New Document missing", doc);
-        
-        assertTrue("file was not created +f",f.exists());
-        System.out.println(f);
+            Object doc = DocumentFactory.getInstance(f.toURI(), null, Level.WARNING);
+            assertNotNull("Document missing", doc);
+
+            Schema s = SchemaFactory.getInstance(new URI("http://mails/refractions/net"));
+
+            path = "mails_out.xml";
+            f = TestData.temp(this, path);
+            if (f.exists()) f.delete();
+            f.createNewFile();
+
+            DocumentWriter.writeDocument(doc, s, f, null);
+
+            doc = DocumentFactory.getInstance(f.toURI(), null, Level.WARNING);
+            assertNotNull("New Document missing", doc);
+
+            assertTrue("file was not created +f", f.exists());
+            System.out.println(f);
         } catch (SAXException e) {
             e.printStackTrace();
             fail(e.toString());
         } catch (Throwable e) {
-        	assertTrue(e instanceof OperationNotSupportedException);
-//            e.printStackTrace();
-//            fail(e.toString()); Operation not supported yet
+            assertTrue(e instanceof OperationNotSupportedException);
+            //            e.printStackTrace();
+            //            fail(e.toString()); Operation not supported yet
         }
     }
 }
